requirements.txt
wxPython >= 4.0.7
mecab-python3 >= 0.996.5
jamdict >= 0.1a6